+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.0 / 2020-04-01
+- [FEATURE] Added --watch option to the build command to watch the src folder for changes and build automatically.
+- [MISC] Moved the commands to src/commands.js to clean up the bin file.
+
 ## 0.1.1 / 2020-03-30
 - [MISC] Added LICENSE file.
 - [MISC] Updated out-of-date dependencies to their latest versions.